requirements.txt
discord.py==2.0.1
ics==0.7.2
pandas==1.5.1
requests==2.27.1
